Improve file and image reading performance by avoiding multiple buffers.
ZnBufferedReadStream>>upToEnd uses a WriteStream which grows the buffer in steps.  For a 4M file approximately 15 buffers (ByteArray) will be created, copied and discarded.  GtBufferedReadStream modifies #upToEnd so that if the underlying stream knows its size, allocate a buffer of the correct size once and fill it.  The underlying stream is checked to confirm it is at the end in case the size is an under-estimate.  One test on a 4M file showed a 7x speed up in reading the file contents, however this will be highly dependent on many aspects of the environment.

ImageReadWriter class>>gtFormFromFileReference: wrapped the stream with 2 ZnPositionableReadStreams, adding another two layers of buffering.  Remove both ZnPositionableReadStream's as the underlying GtBufferedReadStream can be positioned.

As the image processing time is larger than the file reading time the overall performance improvement is about 20%, although this will of course be dependent on the environment.